+++ frontend/src/app/pages/exercises/exercise/shared/simulated-region-overview/simulated-region-overview.module.ts
@@ -45,6 +45,7 @@ import { SimulatedRegionOverviewBehaviorAnswerVehicleRequestsComponent } from '.
 import { RadigoramCardContentResourceRequestComponent } from './radiogram-list/radiogram-card/radigoram-card-content-resource-request/radigoram-card-content-resource-request.component';
 import { SimulatedRegionOverviewBehaviorAutomaticallyDistributeVehiclesComponent } from './tabs/behavior-tab/behaviors/automatically-distribute-vehicles/simulated-region-overview-behavior-automatically-distribute-vehicles.component';
 import { RequestVehiclesComponent } from './tabs/behavior-tab/behaviors/request-vehicles/simulated-region-overview-behavior-request-vehicles.component';
+import { SimulatedRegionOverviewPatientInteractionBarComponent } from './tabs/patients-tab/simulated-region-overview-patient-interaction-bar/simulated-region-overview-patient-interaction-bar.component';
 
 @NgModule({
     declarations: [
@@ -81,6 +82,7 @@ import { RequestVehiclesComponent } from './tabs/behavior-tab/behaviors/request-
         RadigoramCardContentResourceRequestComponent,
         SimulatedRegionOverviewBehaviorAutomaticallyDistributeVehiclesComponent,
         RequestVehiclesComponent,
+        SimulatedRegionOverviewPatientInteractionBarComponent,
     ],
     imports: [
         FormsModule,