+++ shared/src/store/action-reducers/patient.ts
@@ -7,6 +7,10 @@ import {
     PatientStatus,
     patientStatusAllowedValues,
     MapCoordinates,
+    isNotInSimulatedRegion,
+    currentSimulatedRegionIdOf,
+    currentCoordinatesOf,
+    isInSimulatedRegion,
 } from '../../models/utils';
 import {
     changePosition,
@@ -23,6 +27,8 @@ import {
 import { IsLiteralUnion, IsValue } from '../../utils/validators';
 import type { Action, ActionReducer } from '../action-reducer';
 import { ReducerError } from '../reducer-error';
+import { PatientRemovedEvent } from '../../simulation';
+import { sendSimulationEvent } from '../../simulation/events/utils';
 import { updateTreatments } from './utils/calculate-treatments';
 import { getElement } from './utils/get-element';
 import { removeElementPosition } from './utils/spatial-elements';
@@ -53,6 +59,14 @@ export class MovePatientAction implements Action {
     @ValidateNested()
     @Type(() => MapCoordinates)
     public readonly targetPosition!: MapCoordinates;
+}
+
+export class RemovePatientFromSimulatedRegionAction implements Action {
+    @IsValue('[Patient] Remove patient from simulated region' as const)
+    public readonly type = '[Patient] Remove patient from simulated region';
+
+    @IsUUID(4, uuidValidationOptions)
+    public readonly patientId!: UUID;
 }
 
 export class RemovePatientAction implements Action {
@@ -142,9 +156,68 @@ export namespace PatientActionReducers {
         rights: 'participant',
     };
 
+    export const removePatientFromSimulatedRegion: ActionReducer<RemovePatientFromSimulatedRegionAction> =
+        {
+            action: RemovePatientFromSimulatedRegionAction,
+            reducer: (draftState, { patientId }) => {
+                const patient = getElement(draftState, 'patient', patientId);
+
+                if (isNotInSimulatedRegion(patient)) {
+                    throw new ReducerError(
+                        `Patient with Id: ${patientId} was expected to be in simulated region but position was of type: ${patient.position.type}`
+                    );
+                }
+
+                const simulatedRegion = getElement(
+                    draftState,
+                    'simulatedRegion',
+                    currentSimulatedRegionIdOf(patient)
+                );
+                sendSimulationEvent(
+                    simulatedRegion,
+                    PatientRemovedEvent.create(patientId)
+                );
+
+                const coordinates = cloneDeepMutable(
+                    currentCoordinatesOf(simulatedRegion)
+                );
+
+                // place the patient on the right hand side of the simulated region
+
+                coordinates.y -= 0.5 * simulatedRegion.size.height;
+                coordinates.x +=
+                    simulatedRegion.size.width < 0
+                        ? 5
+                        : simulatedRegion.size.width + 5;
+
+                changePositionWithId(
+                    patientId,
+                    MapPosition.create(coordinates),
+                    'patient',
+                    draftState
+                );
+
+                return draftState;
+            },
+            rights: 'trainer',
+        };
+
     export const removePatient: ActionReducer<RemovePatientAction> = {
         action: RemovePatientAction,
         reducer: (draftState, { patientId }) => {
+            const patient = getElement(draftState, 'patient', patientId);
+            if (isInSimulatedRegion(patient)) {
+                const simulatedRegion = getElement(
+                    draftState,
+                    'simulatedRegion',
+                    currentSimulatedRegionIdOf(patient)
+                );
+                sendSimulationEvent(
+                    simulatedRegion,
+                    PatientRemovedEvent.create(patientId)
+                );
+            }
+
             deletePatient(draftState, patientId);
             return draftState;
         },